# backend/skillswap-api-spec.json
{"openapi":"3.1.0","info":{"title":"SkillSwap API","description":"REST API for SkillSwap platform","version":"1.0.0"},"paths":{"/":{"get":{"tags":["Root"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/auth/register":{"post":{"tags":["Authentication","Authentication"],"summary":"Register User","description":"Register a new user","operationId":"register_user_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Register User Api Auth Register Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication","Authentication"],"summary":"Login","description":"Login user with username/email and password","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Login Api Auth Login Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/refresh":{"post":{"tags":["Authentication","Authentication"],"summary":"Refresh Token","description":"Refresh access token using refresh token","operationId":"refresh_token_api_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_refresh_token_api_auth_refresh_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Refresh Token Api Auth Refresh Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/users/":{"get":{"tags":["Users","Users"],"summary":"Get All Users","description":"Get all users (admin only)","operationId":"get_all_users_api_users__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get All Users Api Users  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/users/{user_id}":{"get":{"tags":["Users","Users"],"summary":"Get User","description":"Get user by ID (own user or admin)","operationId":"get_user_api_users__user_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get User Api Users  User Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Users","Users"],"summary":"Update User","description":"Update user by ID (own user or admin)","operationId":"update_user_api_users__user_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Update User Api Users  User Id  Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Users","Users"],"summary":"Delete User","description":"Delete user by ID (admin only)","operationId":"delete_user_api_users__user_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Delete User Api Users  User Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sessions/":{"get":{"tags":["Sessions","Sessions"],"summary":"Get All Sessions","description":"Get all sessions (filtered by user role)","operationId":"get_all_sessions_api_sessions__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get All Sessions Api Sessions  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Sessions","Sessions"],"summary":"Create Session","description":"Create a new session","operationId":"create_session_api_sessions__post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Create Session Api Sessions  Post"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/sessions/{session_id}":{"get":{"tags":["Sessions","Sessions"],"summary":"Get Session","description":"Get a specific session","operationId":"get_session_api_sessions__session_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Session Api Sessions  Session Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Sessions","Sessions"],"summary":"Update Session","description":"Update a session","operationId":"update_session_api_sessions__session_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Update Session Api Sessions  Session Id  Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sessions/{session_id}/start":{"post":{"tags":["Sessions","Sessions"],"summary":"Start Session","description":"Start a session (mentor only)","operationId":"start_session_api_sessions__session_id__start_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Start Session Api Sessions  Session Id  Start Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sessions/{session_id}/end":{"post":{"tags":["Sessions","Sessions"],"summary":"End Session","description":"End a session (mentor only)","operationId":"end_session_api_sessions__session_id__end_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response End Session Api Sessions  Session Id  End Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/webhooks/payment":{"post":{"tags":["Webhooks","Webhooks"],"summary":"Payment Webhook","description":"Webhook for payment events","operationId":"payment_webhook_api_webhooks_payment_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Payment Webhook Api Webhooks Payment Post"}}}}}}},"/api/webhooks/blockchain":{"post":{"tags":["Webhooks","Webhooks"],"summary":"Blockchain Webhook","description":"Webhook for blockchain events","operationId":"blockchain_webhook_api_webhooks_blockchain_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Blockchain Webhook Api Webhooks Blockchain Post"}}}}}}},"/api/blockchain/events/{contract_name}":{"get":{"tags":["Blockchain","Blockchain"],"summary":"Get Events By Contract","description":"Get events by contract name","operationId":"get_events_by_contract_api_blockchain_events__contract_name__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"contract_name","in":"path","required":true,"schema":{"type":"string","title":"Contract Name"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"event_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Event Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Events By Contract Api Blockchain Events  Contract Name  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/blockchain/events/tx/{tx_hash}":{"get":{"tags":["Blockchain","Blockchain"],"summary":"Get Event By Tx Hash","description":"Get event by transaction hash","operationId":"get_event_by_tx_hash_api_blockchain_events_tx__tx_hash__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"tx_hash","in":"path","required":true,"schema":{"type":"string","title":"Tx Hash"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Event By Tx Hash Api Blockchain Events Tx  Tx Hash  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/blockchain/events/address/{address}":{"get":{"tags":["Blockchain","Blockchain"],"summary":"Get Events By Address","description":"Get events for a specific address (as participant)","operationId":"get_events_by_address_api_blockchain_events_address__address__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"address","in":"path","required":true,"schema":{"type":"string","title":"Address"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Events By Address Api Blockchain Events Address  Address  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/blockchain/tokens":{"get":{"tags":["Blockchain","Blockchain"],"summary":"Get Tokens","description":"Get token info","operationId":"get_tokens_api_blockchain_tokens_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Tokens Api Blockchain Tokens Get"}}}}}}},"/api/blockchain/mentors":{"get":{"tags":["Blockchain","Blockchain"],"summary":"Get Blockchain Mentors","description":"Get blockchain mentor list","operationId":"get_blockchain_mentors_api_blockchain_mentors_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Blockchain Mentors Api Blockchain Mentors Get"}}}}}}},"/api/blockchain/transactions":{"post":{"tags":["Blockchain","Blockchain"],"summary":"Create Blockchain Transaction","description":"Create a blockchain transaction","operationId":"create_blockchain_transaction_api_blockchain_transactions_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Create Blockchain Transaction Api Blockchain Transactions Post"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/dashboard":{"get":{"tags":["Admin","Admin"],"summary":"Admin Dashboard","description":"Admin dashboard overview","operationId":"admin_dashboard_api_admin_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Admin Dashboard Api Admin Dashboard Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/db/stats":{"get":{"tags":["Admin","Admin"],"summary":"Database Stats","description":"Database statistics","operationId":"database_stats_api_admin_db_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Database Stats Api Admin Db Stats Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/db/slow-queries":{"get":{"tags":["Admin","Admin"],"summary":"Slow Queries","description":"Analyze slow queries","operationId":"slow_queries_api_admin_db_slow_queries_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Slow Queries Api Admin Db Slow Queries Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/db/backup":{"post":{"tags":["Admin","Admin"],"summary":"Create Backup","description":"Create a database backup","operationId":"create_backup_api_admin_db_backup_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Create Backup Api Admin Db Backup Post"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/db/backups":{"get":{"tags":["Admin","Admin"],"summary":"List Backups","description":"List available backups","operationId":"list_backups_api_admin_db_backups_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response List Backups Api Admin Db Backups Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/db/restore/{backup_name}":{"post":{"tags":["Admin","Admin"],"summary":"Restore Backup","description":"Restore database from backup","operationId":"restore_backup_api_admin_db_restore__backup_name__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"backup_name","in":"path","required":true,"schema":{"type":"string","title":"Backup Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Restore Backup Api Admin Db Restore  Backup Name  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/migrations":{"get":{"tags":["Admin","Admin"],"summary":"List Migrations","description":"List migrations","operationId":"list_migrations_api_admin_migrations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response List Migrations Api Admin Migrations Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/migrations/run":{"post":{"tags":["Admin","Admin"],"summary":"Run Migrations","description":"Run pending migrations","operationId":"run_migrations_api_admin_migrations_run_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Run Migrations Api Admin Migrations Run Post"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/migrations/rollback":{"post":{"tags":["Admin","Admin"],"summary":"Rollback Migration","description":"Rollback migrations","operationId":"rollback_migration_api_admin_migrations_rollback_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"count","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Count"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Rollback Migration Api Admin Migrations Rollback Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_api_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_auth_login_post"},"Body_refresh_token_api_auth_refresh_post":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"Body_refresh_token_api_auth_refresh_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","email","full_name","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/login"}}}}}}